goto文の検討
