+++ programs/index.html
@@ -7,24 +7,69 @@
 </head>
 <body>
     <div id="editor"></div>
-    <button id="run">実行</button>
-    <div id="result" font-size: 18px><p>＜構文解析ログ＞</p></div>
-    <div id="output" font-size: 18px>
-        <p>＜字句解析＞</p>
-        <table id="tokenTable">
-            <thead>
-                <tr>
-                    <th>Line</th>
-                    <th>Type</th>
-                    <th>Value</th>
-                    <th>Number</th>
-                    <th>Level</th>
-                </tr>
-            </thead>
-            <tbody></tbody>
+    <button id="run">run</button>
+    <button id="compile">compile</button>
+    <button id="run_step">step</button>
+    <button id="reset">reset</button>
+    
+    <div id="result">
+        <p>＜構文解析ログ＞</p>
+    </div>
+    
+    <div id="output">
+        <p>＜ステップ実行＞</p>
+    </div>
+    
+    <div id="variable_table_output">
+        <p>＜変数表-ステップ＞</p>
+        <table id="variables_table_output">
+            <tr>
+                <td>変数名</td>
+                <td>値</td>
+            </tr>
+            <tbody>
+            </tbody>
         </table>
     </div>
-    <div id="execute" font-size: 18px><p>＜実行結果＞</p></div>
+
+    <div id="loop">
+        <p>＜ループ実行＞</p>
+    </div>
+
+    <div id="variable_table_loop">
+        <p>＜変数表-ループ＞</p>
+        <table id="variables_table_loop">
+            <tr>
+                <td>変数名</td>
+                <td>値</td>
+            </tr>
+            <tbody>
+            </tbody>
+        </table>
+    </div>
+    
+    <button id="prev_step">&lt;</button>
+    <button id="next_step">&gt;</button>
+    <button id="step_loop">loop</button>
+    <button id="prev_loop">&lt;</button>
+    <button id="next_loop">&gt;</button>
+    
+    <div id="execute">
+        <p>＜実行結果＞</p>
+    </div>
+
+    <div id="variable_table">
+        <p>＜変数表＞</p>
+        <table id="variables_table">
+            <tr>
+                <td>変数名</td>
+                <td>値</td>
+            </tr>
+            <tbody>
+            </tbody>
+        </table>
+    </div>
+    
     <script src="https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.12/ace.js"></script>
     <script src="https://cdn.jsdelivr.net/pyodide/v0.21.3/full/pyodide.js"></script>
     <script src="ace.js"></script>
