+++ programs/scanner.js
@@ -78,112 +78,110 @@ function isFloat(word) {
 }
 
 function isString(word) {
+    // "で囲まれた文字列か'で囲まれた文字列かを判定"
     return /^".*"$/.test(word) || /^'.*'$/.test(word);
 }
 
 function getTokenType(word) {
     if (isInteger(word)) return tk_type.TK_INTEGER;
-    if (isFloat(word)) return tk_type.TK_FLOAT;
-    if (isString(word)) return tk_type.TK_STRING;
-    if (word === 'print') return tk_type.TK_PRINT;
-    if (word === '(') return tk_type.TK_L_PAR;
-    if (word === ')') return tk_type.TK_R_PAR;
-    if (word === ',') return tk_type.TK_COMMA;
-    if (word === '=') return tk_type.TK_EQUAL;
-    if (word === ':') return tk_type.TK_COLON;
-    if (word === '[') return tk_type.TK_L_INDEX;
-    if (word === ']') return tk_type.TK_R_INDEX;
-    if (word === 'for') return tk_type.TK_FOR;
-    if (word === 'in') return tk_type.TK_IN;
-    if (word === 'range') return tk_type.TK_RANGE;
-    if (word === '\t') return tk_type.TK_TAB;
-    if (word === '\n') return tk_type.TK_ENTER;
-    if (word === 'sep') return tk_type.TK_SEPARATOR;
-    if (word === 'end') return tk_type.TK_END;
-    if (word === 'if') return tk_type.TK_IF;
-    if (word === 'elif') return tk_type.TK_ELIF;
-    if (word === 'else') return tk_type.TK_ELSE;
-    if (word === 'while') return tk_type.TK_WHILE;
-    if (word === '{') return tk_type.TK_L_BRACE;
-    if (word === '}') return tk_type.TK_R_BRACE;
-    if (word === '+') return tk_type.TK_PLUS;
-    if (word === '-') return tk_type.TK_MINUS;
-    if (word === '*') return tk_type.TK_MULTIPLY;
-    if (word === '/') return tk_type.TK_DIVIDE;
-    if (word === '>') return tk_type.TK_GREATER;
-    if (word === '<') return tk_type.TK_LESS;
-    if (word === '!') return tk_type.TK_EXCLAMATION;
-    if (word === '#') return tk_type.TK_SHARP;
-    if (word === '%') return tk_type.TK_PERCENT;
-    if (word === 'not') return tk_type.TK_NOT;
-    if (word === '.') return tk_type.TK_DOT;
-    if (word === 'int') return tk_type.TK_INT;
-    if (word === 'float') return tk_type.TK_FL;
-    if (word === 'input') return tk_type.TK_INPUT;
-    if (word === 'map') return tk_type.TK_MAP;
-    if (word === 'list') return tk_type.TK_LIST;
-    if (word === 'split') return tk_type.TK_SPLIT;
-    if (word === 'import') return tk_type.TK_IMPORT;
-    if (word === 'and') return tk_type.TK_AND;
-    if (word === 'or') return tk_type.TK_OR;
-
-    return tk_type.TK_IDENTIFIER;
+    else if (isFloat(word)) return tk_type.TK_FLOAT;
+    else if (word === 'print') return tk_type.TK_PRINT;
+    else if (word === '(') return tk_type.TK_L_PAR;
+    else if (word === ')') return tk_type.TK_R_PAR;
+    else if (word === ',') return tk_type.TK_COMMA;
+    else if (word === '=') return tk_type.TK_EQUAL;
+    else if (word === ':') return tk_type.TK_COLON;
+    else if (word === '[') return tk_type.TK_L_INDEX;
+    else if (word === ']') return tk_type.TK_R_INDEX;
+    else if (word === 'for') return tk_type.TK_FOR;
+    else if (word === 'in') return tk_type.TK_IN;
+    else if (word === 'range') return tk_type.TK_RANGE;
+    else if (word === '\t') return tk_type.TK_TAB;
+    else if (word === '\n') return tk_type.TK_ENTER;
+    else if (word === 'sep') return tk_type.TK_SEPARATOR;
+    else if (word === 'end') return tk_type.TK_END;
+    else if (word === 'if') return tk_type.TK_IF;
+    else if (word === 'elif') return tk_type.TK_ELIF;
+    else if (word === 'else') return tk_type.TK_ELSE;
+    else if (word === 'while') return tk_type.TK_WHILE;
+    else if (word === '{') return tk_type.TK_L_BRACE;
+    else if (word === '}') return tk_type.TK_R_BRACE;
+    else if (word === '+') return tk_type.TK_PLUS;
+    else if (word === '-') return tk_type.TK_MINUS;
+    else if (word === '*') return tk_type.TK_MULTIPLY;
+    else if (word === '/') return tk_type.TK_DIVIDE;
+    else if (word === '>') return tk_type.TK_GREATER;
+    else if (word === '<') return tk_type.TK_LESS;
+    else if (word === '!') return tk_type.TK_EXCLAMATION;
+    else if (word === '#') return tk_type.TK_SHARP;
+    else if (word === '%') return tk_type.TK_PERCENT;
+    else if (word === 'not') return tk_type.TK_NOT;
+    else if (word === '.') return tk_type.TK_DOT;
+    else if (word === 'int') return tk_type.TK_INT;
+    else if (word === 'float') return tk_type.TK_FL;
+    else if (word === 'input') return tk_type.TK_INPUT;
+    else if (word === 'map') return tk_type.TK_MAP;
+    else if (word === 'list') return tk_type.TK_LIST;
+    else if (word === 'split') return tk_type.TK_SPLIT;
+    else if (word === 'import') return tk_type.TK_IMPORT;
+    else if (word === 'and') return tk_type.TK_AND;
+    else if (word === 'or') return tk_type.TK_OR;
+    else if (isIdentifier(word)) return tk_type.TK_IDENTIFIER;
+    else if (isString(word)) return tk_type.TK_STRING;
 }
 
 function tokenNumber(word) {
     if (isInteger(word)) return 2;
-    if (isFloat(word)) return 3;
-    if (isString(word)) return 4;
-    if (word === 'print') return 5;
-    if (word === '(') return 6;
-    if (word === ')') return 7;
-    if (word === ',') return 8;
-    if (word === '=') return 9;
-    if (word === ':') return 10;
-    if (word === '[') return 11;
-    if (word === ']') return 12;
-    if (word === 'for') return 13;
-    if (word === 'in') return 14;
-    if (word === 'range') return 15;
-    if (word === '\t' || word === '    ') return 16;
-    if (word === '\n') return 17;
-    if (word === 'sep') return 18;
-    if (word === 'end') return 19;
-    if (word === 'if') return 20;
-    if (word === 'elif') return 21;
-    if (word === 'else') return 22;
-    if (word === 'while') return 23;
-    if (word === '{') return 24;
-    if (word === '}') return 25;
-    if (word === '+') return 26;
-    if (word === '-') return 27;
-    if (word === '*') return 28;
-    if (word === '/') return 29;
-    if (word === '>') return 30;
-    if (word === '<') return 31;
-    if (word === '!') return 32;
-    if (word === '#') return 33;
-    if (word === '%') return 34;
-    if (word === 'not') return 35;
-    if (word === '.') return 36;
-    if (word === 'int') return 37;
-    if (word === 'float') return 38;
-    if (word === 'input') return 39;
-    if (word === 'map') return 40;
-    if (word === 'list') return 41;
-    if (word === 'split') return 42;
-    if (word === 'import') return 43;
-    if (word === 'and') return 44;
-    if (word === 'or') return 45;
-    return 1;
+    else if (isFloat(word)) return 3;
+    else if (word === 'print') return 5;
+    else if (word === '(') return 6;
+    else if (word === ')') return 7;
+    else if (word === ',') return 8;
+    else if (word === '=') return 9;
+    else if (word === ':') return 10;
+    else if (word === '[') return 11;
+    else if (word === ']') return 12;
+    else if (word === 'for') return 13;
+    else if (word === 'in') return 14;
+    else if (word === 'range') return 15;
+    else if (word === '\t' || word === '    ') return 16;
+    else if (word === '\n') return 17;
+    else if (word === 'sep') return 18;
+    else if (word === 'end') return 19;
+    else if (word === 'if') return 20;
+    else if (word === 'elif') return 21;
+    else if (word === 'else') return 22;
+    else if (word === 'while') return 23;
+    else if (word === '{') return 24;
+    else if (word === '}') return 25;
+    else if (word === '+') return 26;
+    else if (word === '-') return 27;
+    else if (word === '*') return 28;
+    else if (word === '/') return 29;
+    else if (word === '>') return 30;
+    else if (word === '<') return 31;
+    else if (word === '!') return 32;
+    else if (word === '#') return 33;
+    else if (word === '%') return 34;
+    else if (word === 'not') return 35;
+    else if (word === '.') return 36;
+    else if (word === 'int') return 37;
+    else if (word === 'float') return 38;
+    else if (word === 'input') return 39;
+    else if (word === 'map') return 40;
+    else if (word === 'list') return 41;
+    else if (word === 'split') return 42;
+    else if (word === 'import') return 43;
+    else if (word === 'and') return 44;
+    else if (word === 'or') return 45;
+    else if (isIdentifier(word)) return 1;
+    else if (isString(word)) return 4;
 }
 
 // コードを解析する関数
 function processCode(code) {
     let lines = code.split('\n');
     let lineNumber = 1;
-    let tableBody = document.getElementById('tokenTable').getElementsByTagName('tbody')[0];
-    tableBody.innerHTML = ''; // テーブルの内容をクリア
 
     // 解析処理
     lines.forEach(line => {
@@ -198,16 +196,16 @@ function processCode(code) {
             if (/\s{4,}/.test(word) && word.length % 4 === 0) {
                 for (let j = 0; j < tabCount; j++) {
                     let newToken = createToken(/\t/, tk_type.TK_TAB, lineNumber, tokenNumber('\t'), tabCount);
-                    addRowToTable(newToken);
                     parsedTokens.push(newToken);
+                    //console.log(newToken);
                 }
             } else {
                 if (word === '"' || word === "'") {
                     if (isString) {
                         combinedString += word;
                         let newToken = createToken(combinedString, tk_type.TK_STRING, lineNumber, tokenNumber(combinedString), tabCount);
-                        addRowToTable(newToken);
                         parsedTokens.push(newToken);
+                        //console.log(newToken);
                         combinedString = '';
                         isString = false;
                     } else {
@@ -225,8 +223,8 @@ function processCode(code) {
                             i++;
                         }
                         let newToken = createToken(floatValue, tk_type.TK_FLOAT, lineNumber, tokenNumber(floatValue), tabCount);
-                        addRowToTable(newToken);
                         parsedTokens.push(newToken);
+                        //console.log(newToken);
                     } else if (word === '.' && i + 1 < words.length && isInteger(words[i + 1])) {
                         let floatValue = word;
                         if (i + 1 < words.length && isInteger(words[i + 1])) {
@@ -234,44 +232,29 @@ function processCode(code) {
                             i++;
                         }
                         let newToken = createToken(floatValue, tk_type.TK_FLOAT, lineNumber, tokenNumber(floatValue), tabCount);
-                        addRowToTable(newToken);
                         parsedTokens.push(newToken);
+                        //console.log(newToken);
                     } else if (isInteger(word)) {
                         let newToken = createToken(word, tk_type.TK_INTEGER, lineNumber, tokenNumber(word), tabCount);
-                        addRowToTable(newToken);
+
                         parsedTokens.push(newToken);
+                        //console.log(newToken);
                     } else {
                         let type = getTokenType(word);
                         let newToken = createToken(word, type, lineNumber, tokenNumber(word), tabCount);
-                        addRowToTable(newToken);
                         parsedTokens.push(newToken);
+                        //console.log(newToken);
                     }
                 }
             }
         }
 
         let newToken = createToken(/\n/, tk_type.TK_ENTER, lineNumber, tokenNumber('\n'), tabCount);
-        addRowToTable(newToken);
         parsedTokens.push(newToken);
+        //console.log(newToken);
         lineNumber++;
     });
     return parsedTokens;
-}
-
-// テーブルに行を追加する関数
-function addRowToTable(token) {
-    let tableBody = document.getElementById('tokenTable').getElementsByTagName('tbody')[0];
-    let newRow = tableBody.insertRow();
-    let lineCell = newRow.insertCell(0);
-    let typeCell = newRow.insertCell(1);
-    let valueCell = newRow.insertCell(2);
-    let numberCell = newRow.insertCell(3);
-    let tabCountCell = newRow.insertCell(4);
-    lineCell.innerHTML = token.lineNumber;
-    typeCell.innerHTML = token.type;
-    valueCell.innerHTML = token.word;
-    numberCell.innerHTML = token.tkNumber;
-    tabCountCell.innerHTML = token.tabCount;
 }
 
 // 実行ボタンをクリックしたときの処理
