--- renew/eval.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-<head>
-    <meta charset="UTF-8">
-    <title>eval</title>
-</head>
-<body>
-    <script src="https://cdn.jsdelivr.net/pyodide/v0.21.3/full/pyodide.js"></script>
-    <script>
-        async function runPythonCode() {
-            // Pyodideをロード
-            let pyodide = await loadPyodide();
-            // Pythonコードを実行
-            pyodide.runPython(`
-                x = [[1, 2, 3], [4, 5, 6]]
-                for rows in x:
-                    for columns in rows:
-                        print(columns)
-            `);
-        }
-        runPythonCode();
-    </script>
-</body>
-</html>
