+++ renew/styles.css
@@ -11,16 +11,16 @@ body {
 /* エディタ部分の配置 */
 #editor {
     position: absolute;
-    top: 0;
+    bottom: 20%;
     left: 0;
     width: 50%;
     height: 50%;
-    border: 1px solid #ccc;
+    border: 2px solid #ccc;
 }
 
 #compile {
     position: absolute;
-    bottom: 44.5vh;
+    bottom: 15.3%;
     left: 0;
     width: 12%;
     height: 5%;
@@ -31,9 +31,9 @@ body {
 
 #run {
     position: absolute;
-    bottom: 44.5vh;
-    left: 12%;
-    width: 12%;
+    top : 29%;
+    right: 41%;
+    width: 8%;
     height: 5%;
     border: 2px solid #0a0a0a;
     overflow-y: auto;
@@ -42,8 +42,8 @@ body {
 
 #reset {
     position: absolute;
-    bottom: 44.5vh;
-    left: 24%;
+    bottom: 15.3%;
+    left: 12%;
     width: 12%;
     height: 5%;
     border: 2px solid #0a0a0a;
@@ -54,11 +54,11 @@ body {
 /* 構文解析ログの配置 */
 #result {
     position: absolute;
-    bottom: 15vh;
+    bottom: 0;
     left: 0;
     width: 50%;
-    height: 25%;
-    border: 1px solid #080cf1;
+    height: 15%;
+    border: 2px solid #080cf1;
     overflow-y: auto;
     font-size: 18px;
 }
@@ -66,29 +66,29 @@ body {
 
 #output {
     position: absolute;
-    top: 0;
-    right: 24%;
-    width: 24%;
-    height: 24%;
-    border: 1px solid #08f02e;
+    bottom: 35%;
+    right: 0;
+    width: 48%;
+    height: 30%;
+    border: 4px solid #f04a08;
     overflow-y: auto;
     font-size: 18px;
 }
 
 #variable_table_output {
     position: absolute;
-    top: 24%;
-    right: 24%;
-    width: 24%;
-    height: 24%;
-    border: 1px solid #08f02e;
+    bottom: 5%;
+    right: 0%;
+    width: 48%;
+    height: 30%;
+    border: 4px solid #f04a08;
     overflow-y: auto;
     font-size: 18px;
 }
 
 #run_step {
     position: absolute;
-    bottom: 46%;
+    top : 29%;
     right: 41%;
     width: 8%;
     height: 5%;
@@ -99,7 +99,7 @@ body {
 
 #prev_step {
     position: absolute;
-    bottom: 46%;
+    top : 29%;
     right: 33%;
     width: 8%;
     height: 5%;
@@ -110,7 +110,7 @@ body {
 
 #next_step {
     position: absolute;
-    bottom: 46%;
+    top : 29%;
     right: 25%;
     width: 8%;
     height: 5%;
@@ -121,30 +121,30 @@ body {
 
 #loop {
     position: absolute;
-    top: 0;
+    bottom: 35%;
     right: 0;
-    width: 24%;
-    height: 24%;
-    border: 1px solid #ecf008;
+    width: 48%;
+    height: 30%;
+    border: 4px solid #ecf008;
     overflow-y: auto;
     font-size: 18px;
 }
 
 #variable_table_loop {
     position: absolute;
-    top: 24%;
-    right: 0;
-    width: 24%;
-    height: 24%;
-    border: 1px solid #ecf008;
+    bottom: 5%;
+    right: 0%;
+    width: 48%;
+    height: 30%;
+    border: 4px solid #ecf008;
     overflow-y: auto;
     font-size: 18px;
 }
 
 #step_loop {
     position: absolute;
-    bottom: 46%;
-    right: 16%;
+    top : 29%;
+    right: 41%;
     width: 8%;
     height: 5%;
     border: 2px solid #0a0a0a;
@@ -154,8 +154,8 @@ body {
 
 #prev_loop {
     position: absolute;
-    bottom: 46%;
-    right: 8%;
+    top : 29%;
+    right: 33%;
     width: 8%;
     height: 5%;
     border: 2px solid #0a0a0a;
@@ -165,8 +165,8 @@ body {
 
 #next_loop {
     position: absolute;
-    bottom: 46%;
-    right: 0;
+    top : 29%;
+    right: 25%;
     width: 8%;
     height: 5%;
     border: 2px solid #0a0a0a;
@@ -176,22 +176,22 @@ body {
 
 #execute {
     position: absolute;
-    bottom: 15vh;
-    right: 24%;
-    width: 24%;
+    bottom : 35%;
+    right: 0;
+    width: 48%;
     height: 30%;
-    border: 1px solid #f04a08;
+    border: 4px solid #08f070;
     overflow-y: auto;
     font-size: 18px;
 }
 
 #variable_table {
     position: absolute;
-    bottom: 15vh;
+    bottom : 5%;
     right: 0%;
-    width: 24%;
+    width: 48%;
     height: 30%;
-    border: 1px solid #f04a08;
+    border: 4px solid #08f070;
     overflow-y: auto;
     font-size: 18px;
 }
@@ -204,7 +204,7 @@ body {
 #variables_table th, #variables_table td, #variables_table tbody th, #variables_table tbody td {
     padding: 6px;
     text-align: left;
-    border: 1px solid #ddd;
+    border: 4px solid #ddd;
 }
 
 #variables_table_output {
@@ -215,7 +215,7 @@ body {
 #variables_table_output th, #variables_table_output td, #variables_table_output tbody th, #variables_table_output tbody td {
     padding: 6px;
     text-align: left;
-    border: 1px solid #ddd;
+    border: 4px solid #ddd;
 }
 
 #variables_table_loop {
@@ -226,9 +226,57 @@ body {
 #variables_table_loop th, #variables_table_loop td, #variables_table_loop tbody th, #variables_table_loop tbody td {
     padding: 6px;
     text-align: left;
-    border: 1px solid #ddd;
+    border: 4px solid #ddd;
 }
 
 .highlight-line {
     background-color: rgb(255, 81, 0) !important;
+}
+
+.hidden {
+    display: none;
+}
+
+#show_execute {
+    position: absolute;
+    bottom: 0;
+    right: 32%;
+    width: 16%;
+    height: 5%;
+    border: 4px solid #08f070;
+    overflow-y: auto;
+    font-size: 18px;
+}
+
+#show_step {
+    position: absolute;
+    bottom: 0;
+    right: 16%;
+    width: 16%;
+    height: 5%;
+    border: 4px solid #f04a08;
+    overflow-y: auto;
+    font-size: 18px;
+}
+
+#show_loop {
+    position: absolute;
+    bottom: 0;
+    right: 0;
+    width: 16%;
+    height: 5%;
+    border: 4px solid #ecf008;
+    overflow-y: auto;
+    font-size: 18px;
+}
+
+#mondai {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 98%;
+    height: 25%;
+    border: 4px solid #0e0d0e;
+    overflow-y: auto;
+    font-size: 18px;
 }
